% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combData.R
\name{combData}
\alias{combData}
\title{Title Combine reference and user's sample data sets together}
\usage{
combData(
  dataset,
  reference,
  class = c("rgset", "betas"),
  normType = c("Noob", "Funnorm", "Quantile", "Quantile.b", "none"),
  cellTypes = c("Bas", "Bmem", "Bnv", "CD4mem", "CD4nv", "CD8mem", "CD8nv", "Eos",
    "Mono", "Neu", "NK", "Treg"),
  verbose = T
)
}
\arguments{
\item{dataset}{An RGchannelSet or a beta matrix of the users' DNA methylation data set}

\item{reference}{An RGchannelSet of the reference DNA methylation data set, preferentially the output from the \code{getRef()} function, to ensure the \code{pData()} formatting is appropriate}

\item{class}{A character specifying the argument \code{dataset}'s class, with the option or rgset (for RGChannelSet) or betas (for beta matrix)}

\item{normType}{A character specifying the method for normalizing user sample and reference data sets together, with the options of "Noob", "Funnorm", "Quantile", "Quantile.b", and "none" -- refer to the Details section for more information on the options.}

\item{cellTypes}{A vector of characters specifying the cell types to be estimated. If the \code{getRef()} function was used, use the output as shown in the Example.}

\item{verbose}{A Boolean, specifying whether the function should be verbose}
}
\value{
A beta matrix of
}
\description{
Title Combine reference and user's sample data sets together
}
\examples{
# Load example blood cell mixture, subsetted from the IDOL dataset (GSE110554)
test_dat <- IDOL_mixed_cells
# Obtain reference data set with the `getRef()` function
ref_dat <- getRef(ref = "IDOL", normType = "None")
# Combine sample and reference data sets together, followed by normalization (if selected)
comb_dat <- combData(dataset = test_dat, reference = ref_dat$reference, class = "rgset", normType = "none", cellTypes = ref_dat$cellTypes)
}
