% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRef.R
\name{getRef}
\alias{getRef}
\title{Extract the correct reference data set}
\usage{
getRef(
  ref = c("Reinius", "IDOL", "IDOL_extended", "Mixed", "Cord", "DLPFC", "Middleton"),
  normType = c("None", "Noob", "Funnorm", "Quantile")
)
}
\arguments{
\item{ref}{A character representing the reference dataset, with the options of
"Reinius", "IDOL", "IDOL_extended", "Mixed", "Cord", "DLPFC", and "Middleton"}

\item{normType}{a character representing the normalization method, with the
options of "None", "Noob", "Funnorm", "Quantile"}
}
\value{
A list of the requested reference data set (RGchannelSet Object) and
the cell types of the sample (vector of character)
}
\description{
Extract the correct reference data set
}
\examples{
# Request the IDOL reference (2016) without normalization
getRef(ref = "IDOL", normType = "None")
}
