% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identClust.R
\name{identClust}
\alias{identClust}
\title{Calculate the stability of selected features with pvClust}
\usage{
identClust(dataNormed, probes, parallel = TRUE)
}
\arguments{
\item{dataNormed}{A list of dataframe containing the normalized data, output of \code{combData()}}

\item{probes}{A list of probes to perform cell type prediction with and their corresponding coefficient, output of \code{pickProbes()}}

\item{parallel}{A SOCKcluster object specifying the number of processes to parallelize, created with \code{doParallel::makeCluster()}}
}
\value{
Approximately Unbiased (AU) measure of clusters based on hierarchical clustering results, representing the stability of each clusters via multiscale bootstrap resampling
}
\description{
Calculate the stability of selected features with pvClust
}
\examples{
# Load example blood cell mixture, subsetted from the IDOL dataset (GSE110554)
test_dat <- CellsPickMe::IDOL_mixed_cells
# Obtain reference data set with the `getRef()` function
ref_dat <- getRef(ref = "IDOL", normType = "None")
# Combine sample and reference data sets together, followed by normalization (if selected)
comb_dat <- combData(dataset = test_dat,
reference = ref_dat$reference, class = "rgset", normType = "None", cellTypes = ref_dat$cellTypes)
# Pick probes with repeated cross validation with T-test
probes <- pickProbes(dataNormed = comb_dat, probeList = "Ttest", probeSelect = "both", nProbes = 100, min.delta.beta = 0.05)
# Create parallelization clusters and calculate cluster stability
clustAU <- identClust(dataNormed = comb_dat, probes = probes, parallel = TRUE)
}
