% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idol.a_450}
\alias{idol.a_450}
\title{IDOL probes for 450k array - adult blood}
\format{
\subsection{}{

A vector of characters of length 350
The data is obtained from a publication by Salas et al. Please cite the following if using this data.
Salas LA, Koestler DC, Butler RA, Hansen HM et al. An optimized library for reference-based deconvolution of whole-blood biospecimens assayed using the Illumina HumanMethylationEPIC BeadArray. Genome Biol 2018 May 29;19(1):64. PMID: 29843789
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29843789/}
}
\usage{
idol.a_450
}
\description{
CpG names of probes selected by IDOL, for 450K DNAm microarray
}
\keyword{datasets}
