% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idol.a_ext}
\alias{idol.a_ext}
\title{IDOL probes for EPIC array - extended 12 cell type reference}
\format{
\subsection{}{

A vector of characters of length 1200
The data is obtained from a publication by Salas et al. Please cite the following if using this data.
Salas LA, Zhang Z, Koestler DC, Butler RA et al. Enhanced cell deconvolution of peripheral blood using DNA methylation for high-resolution immune profiling. Nat Commun 2022 Feb 9;13(1):761. PMID: 35140201
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/35140201/}
}
\usage{
idol.a_ext
}
\description{
CpG names of probes selected by IDOL, for the extended blood cell type reference
}
\keyword{datasets}
