% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idol.c}
\alias{idol.c}
\title{IDOL probes for 450k array - cord blood}
\format{
\subsection{}{

A vector of characters of length 517
The data is obtained from a publication by Gervin et al.. Please cite the following if using this data.
Gervin, K., Salas, L. A., Bakulski, K. M., Van Zelm, M. C., Koestler, D. C., Wiencke, J. K., ... & Jones, M. J. (2019). Systematic evaluation and validation of reference and library selection methods for deconvolution of cord blood DNA methylation data. Clinical epigenetics, 11(1), 1-15.
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29843789/}
}
\usage{
idol.c
}
\description{
CpG names of probes selected by IDOL, for cord blood cell type reference
}
\keyword{datasets}
