% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDOL_mixed_cells}
\alias{IDOL_mixed_cells}
\title{IDOL mixed blood cell DNA methylation}
\format{
\subsection{}{

An RGChannelSet with 1,051,815 rows and 12 columns
This is a DNA methylation data set of artificially combined blood cell mixture, as measured on Illumina's EPIC microarray. The data represents the averaged DNA methylation level of the blood cell mixture.
The data is extracted from GSE110554, a dataset created and published by Salas et al. Please cite the following if using this data.
Salas LA, Koestler DC, Butler RA, Hansen HM et al. An optimized library for reference-based deconvolution of whole-blood biospecimens assayed using the Illumina HumanMethylationEPIC BeadArray. Genome Biol 2018 May 29;19(1):64. PMID: 29843789
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE110554}
}
\usage{
IDOL_mixed_cells
}
\description{
A subset of data from GSE110554
}
\keyword{datasets}
