% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Middleton_pd}
\alias{Middleton_pd}
\title{Middleton et al. Saliva reference pheno data}
\format{
\subsection{}{

A dataframe of 60 rows and 1 column
This data contains the cell type label of the Middleton saliva reference, allowing for usability of the \code{getRef()} function when Middleton reference is called.
The data is extracted from GSE147318, a dataset created and published by Middleton et al. Please cite the following if using this data.
Middleton LYM, Dou J, Fisher J, Heiss JA et al. Saliva cell type DNA methylation reference panel for epidemiological studies in children. Epigenetics 2022 Jan-Feb;17(2):161-177. PMID: 33588693
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE147318}
}
\usage{
Middleton_pd
}
\description{
The phenotype data of GSE147318
}
\keyword{datasets}
