% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickCompProbesCaret.R
\name{pickCompProbesCaret}
\alias{pickCompProbesCaret}
\title{Pick features for cell type prediction with caret by repeated cross validation}
\usage{
pickCompProbesCaret(
  betas,
  meta,
  ct,
  ps,
  min.delta.beta,
  p.val,
  caretMods,
  filterK = 1000,
  seed = 1234,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{betas}{A beta matrix of reference DNA methylation data that will be used to select for features for cell type prediction}

\item{meta}{A data frame of phenotype data, specifying which of the reference samples are of what cell types}

\item{ct}{A vector of characters specifying the cell types to deconvolute}

\item{ps}{A character of either "any" or "both" to specify first line filter for T test prior to passing to machine learning algorithms}

\item{min.delta.beta}{A numeric variable defining T test minimum delta beta for first line filter prior to passing to machine learning algorithms}

\item{p.val}{A numeric variable defining maximum T test p.value for first line filter prior to passing to machine learning algorithms}

\item{caretMods}{A vector of characters, selecting the models to use to pick the cell type prediction features, options include "lasso", "EL", "BLR", "CART", "RF", "GBM", "PLDA", "GAnRF", "GAnNB", "GAnSVM", and "GAnNN"}

\item{filterK}{An integer, representing the number probes to input to the machine learning algorithms with top T test probes}

\item{seed}{An integer specifying the seed for reproducibility}

\item{plot}{A Boolean specifying whether to plot a heatmap showing the clustering performance using the probes selected}

\item{verbose}{A Boolean specifying whether the function should be verbose or not}
}
\description{
Pick features for cell type prediction with caret by repeated cross validation
}
