% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCT.R
\name{predictCT}
\alias{predictCT}
\title{Estimate cell type proportion based on sample DNAm data and coefficients obtained from the reference data}
\usage{
predictCT(
  dataNormed,
  probes,
  method,
  conditions = NULL,
  removenRBC = F,
  verbose = TRUE,
  cetygo = TRUE
)
}
\arguments{
\item{dataNormed}{A list of dataframe containing the normalized data, output of \code{combData()}}

\item{probes}{A list of probes to perform cell type prediction with and their corresponding coefficient, output of \code{pickProbes()}}

\item{method}{A character specifying the regression method. Options include "CP", "RPC", and "SVR"}

\item{conditions}{A character specifying specific conditions for model fitting. Default is NULL}

\item{removenRBC}{A Boolean specifying whether nucleated red blood cell (nRBC) proportion should be estimated, if using a reference with nRBC}

\item{verbose}{A Boolean specifying whether the function should be verbose or not}

\item{cetygo}{A Boolean specifying whether the CETYGO score should be calculated to estimate reference appropriateness}
}
\value{
A matrix containing the estimated cell type proportion for the user samples
}
\description{
Estimate cell type proportion based on sample DNAm data and coefficients obtained from the reference data
}
\examples{
# Load example blood cell mixture, subsetted from the IDOL dataset (GSE110554)
test_dat <- CellsPickMe::IDOL_mixed_cells
# Obtain reference data set with the `getRef()` function
ref_dat <- getRef(ref = "IDOL", normType = "None")
# Combine sample and reference data sets together, followed by normalization (if selected)
comb_dat <- combData(dataset = test_dat, reference = ref_dat$reference, class = "rgset", normType = "None", cellTypes = ref_dat$cellTypes)
# Pick probes with repeated cross validation with lasso and elastic net
probes <- probes <- pickProbes(dataNormed = comb_dat, probeList = "Ttest", probeSelect = "both", nProbes = 100, min.delta.beta = 0.05)
# Estimate cell type proportion
out <- predictCT(dataNormed = comb_dat, probes = probes, method = "CP", cetygo = TRUE)
}
